ai_chat/_test/sql/sql-local-swarm-agent/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp>=3.10.0
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
certifi==2024.8.30
cfgv==3.4.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
distlib==0.3.9
distro==1.9.0
docstring_parser==0.16
filelock==3.16.1
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
identify==2.6.1
idna==3.10
iniconfig==2.0.0
instructor==1.5.2
jiter==0.5.0
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.1.0
nodeenv==1.9.1
numpy==2.1.2
openai>=1.0.0
google-generativeai>=0.3.0
packaging==24.1
platformdirs==4.3.6
pluggy==1.5.0
pre_commit==4.0.1
propcache==0.2.0
pydantic>=2.9.0
pydantic_core==2.23.4
Pygments==2.18.0
pytest==8.3.3
python-dotenv>=1.0.0
PyYAML==6.0.2
requests==2.32.3
rich>=13.9.0
shellingham==1.5.4
sniffio==1.3.1
swarm-ai>=0.1.0
tenacity==8.5.0
tqdm==4.66.5
typer>=0.12.0
typing_extensions==4.12.2
urllib3==2.2.3
virtualenv==20.26.6
yarl==1.15.2
